Removed blobs from files.  Deleted Login Folder, and moved components into Root of '/components'
