+++ src/app.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Router } from 'react-native-router-flux';
-import scenes from './routes/index.js';
+import scenes from './routes';
 
 class App extends Component {
   render() {
