+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import routes from './routes.js';
-import form from './Form/index.js';
+import routes from './routes';
+import form from './Form';
 // ... other reducers
 
 export default combineReducers({
