+++ src/routes/Login.js
@@ -1,15 +1,13 @@
 import React, { Component } from 'react';
 import {
-  View,
-  Text
+  View
 } from 'react-native';
 import {
   Actions
 } from 'react-native-router-flux';
-import TextField from '../components/Login/TextFields.js';
-import Password from '../components/Login/TextFieldPassword.js';
-import Button from '../components/Button.js';
-
+import TextField from '../components/TextFields';
+import Password from '../components/TextFieldPassword';
+import Button from '../components/Button';
 
 class Login extends Component {
 
@@ -26,7 +24,7 @@ class Login extends Component {
   render() {
     return (
       <View>
-        <Text>Login</Text>
+        <TextField />
         <View>
           <Password />
         </View>
